requirements.txt
absl-py==0.15.0
appdirs==1.4.4
astunparse==1.6.3
audeer==1.16.0
audformat==0.11.6
audinterface==0.7.0
audiofile==1.0.1
audioread==2.1.9
audobject==0.5.1
audresample==0.1.6
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
clang==5.0
click==8.0.3
colorama==0.4.4
cycler==0.10.0
decorator==5.1.0
envyaml==1.9.210927
ffmpeg==1.4
filelock==3.3.2
flatbuffers==1.12
fonttools==4.25.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.1.0
huggingface-hub==0.1.0
idna==3.3
iso-639==0.4.5
joblib==1.1.0
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282420641/work
librosa==0.8.0
llvmlite==0.37.0
Markdown==3.3.4
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1634667019719/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
munkres==1.1.4
numba==0.54.1
numpy==1.19.5
oauthlib==3.1.1
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opensmile==2.2.0
opt-einsum==3.3.0
oyaml==1.0
packaging==21.2
pandas==1.2.5
Pillow==8.4.0
pooch==1.5.2
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-speech-features==0.6
pytz==2021.3
PyYAML==6.0
regex==2021.11.2
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.1
sip==4.19.13
six==1.15.0
SoundFile==0.10.3.post1
sox==1.4.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.6.0
tensorflow-hub==0.12.0
termcolor==1.1.0
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.10.0
torchaudio==0.10.0
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
tqdm==4.62.3
transformers==4.12.3
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
